Move legacy specs to their respective new spec files and delete the old specfile (from the wrong location)
