+++ spec/api/chemotion/attachment_api_spec.rb
@@ -281,4 +281,131 @@ describe Chemotion::AttachmentAPI do
   describe 'GET /api/v1/attachments/svgs' do
     pending 'not yet implemented'
   end
+
+  # TODO: Check these specs and remove everything that is already covered by the specs above
+  #       Refactor the rest to match the spec structure as shown above
+  context 'legacy specs from spec/api/attachment_api_spec.rb' do
+    let(:file_upload) do
+      {
+        file_1: fixture_file_upload(Rails.root.join('spec/fixtures/upload.txt'), 'text/plain'),
+        file_2: fixture_file_upload(Rails.root.join('spec/fixtures/upload.txt'), 'text/plain')
+      }
+    end
+    let(:img_upload) do
+      {
+        file_1: fixture_file_upload(Rails.root.join('spec/fixtures/upload.jpg'))
+      }
+    end
+
+    let(:user) { create(:user, first_name: 'Person', last_name: 'Test') }
+    let(:u2) { create(:user) }
+    let(:group) { create(:group) }
+    let(:c1) { create(:collection, user_id: user.id) }
+    let!(:cont_s1_root) { create(:container) }
+    let!(:s1) do
+      create(:sample_without_analysis, name: 'sample 1', container: cont_s1_root)
+    end
+    let!(:cont_s1_analyses) { create(:container, container_type: 'analyses') }
+    let!(:cont_s1_analysis) { create(:analysis_container) }
+    let!(:new_attachment) do
+      create(
+        :attachment,
+        storage: 'tmp', key: '8580a8d0-4b83-11e7-afc4-85a98b9d0194',
+        filename: 'upload.jpg',
+        file_path: File.join(Rails.root, 'spec/fixtures/upload.jpg'),
+        created_by: user.id, created_for: user.id
+      )
+    end
+    let(:new_local_attachment) { build(:attachment, storage: 'local') }
+
+    context 'authorized user logged in' do
+      let(:attachments) do
+        Attachment.where(created_by: user, filename: 'upload.txt')
+      end
+      let(:img_attachments) do
+        Attachment.where(created_by: user, filename: 'upload.jpg')
+      end
+
+      before do
+        allow_any_instance_of(WardenAuthentication).to receive(:current_user)
+          .and_return(user)
+
+        CollectionsSample.create!(sample: s1, collection: c1)
+
+        cont_s1_root.children << cont_s1_analyses
+        cont_s1_root.save!
+        cont_s1_analyses.children << cont_s1_analysis
+        cont_s1_analyses.save!
+
+        img_attachments.last.update!(
+          attachable_id: cont_s1_analysis.id,
+          attachable_type: 'Container'
+        )
+      end
+
+      after(:all) do
+        `rm -rf #{File.join(Rails.root, 'tmp', 'test')}`
+        puts "delete tmp folder #{File.join(Rails.root, 'tmp', 'test')} "
+      end
+
+      describe 'upload files thru POST attachments/upload_dataset_attachments' do
+        before do
+          post '/api/v1/attachments/upload_dataset_attachments', params: file_upload
+        end
+
+        it 'creates attachments for each file' do
+          expect(attachments.count).to eq 2
+        end
+
+        it 'stores file localy' do
+          expect(File.exist?(attachments.last.store.path)).to be true
+        end
+      end
+
+      describe 'upload img thru POST attachments/upload_dataset_attachments' do
+        before do
+          post '/api/v1/attachments/upload_dataset_attachments', params: img_upload
+          img_attachments.reload.last.update!(
+            attachable_id: cont_s1_analysis.id,
+            attachable_type: 'Container'
+          )
+        end
+
+        it 'creates attachments for each file' do
+          expect(img_attachments.count).to eq 2
+        end
+
+        it 'stores file localy' do
+          expect(File.exist?(img_attachments.last.store.path)).to be true
+        end
+
+        it 'creates thumbnail localy' do
+          expect(File.exist?(img_attachments.last.store.thumb_path)).to be true
+        end
+
+        describe 'Return Base64 encoded thumbnail' do
+          before do
+            get "/api/v1/attachments/thumbnail/#{img_attachments.last.id}"
+          end
+
+          it 'creates attachments for each file' do
+            encoded_thumbnail = Base64.encode64(img_attachments.last.read_thumbnail)
+            expect(response.body).to include(encoded_thumbnail.inspect)
+          end
+        end
+
+        describe 'Return Base64 encoded thumbnails' do
+          before do
+            params = { ids: [img_attachments.reload.last.id] }
+            post '/api/v1/attachments/thumbnails', params: params
+          end
+
+          it 'creates attachments for each file' do
+            encoded_thumbnail = Base64.encode64(img_attachments.last.read_thumbnail)
+            expect(response.body).to include(encoded_thumbnail.inspect)
+          end
+        end
+      end
+    end
+  end
 end
